++tree support for move branch

+++ sqlalchemy_mptt/mixins.py
@@ -18,8 +18,8 @@ from sqlalchemy.orm import backref, relationship
 
 class NestedSetsExtension(object):
 
-    @staticmethod
-    def mptt_before_insert(mapper, connection, instance):
+    @classmethod
+    def mptt_before_insert(self, mapper, connection, instance):
         if not instance.parent_id:
             instance.left = 1
             instance.right = 2
@@ -57,8 +57,8 @@ class NestedSetsExtension(object):
             instance.left = right_most_sibling
             instance.right = right_most_sibling + 1
 
-    @staticmethod
-    def mptt_after_delete(mapper, connection, instance):
+    @classmethod
+    def mptt_after_delete(self, mapper, connection, instance, delete=True):
 
         table = mapper.mapped_table
         lft = instance.left
@@ -68,10 +68,11 @@ class NestedSetsExtension(object):
 
         # Delete node or baranch of node
         # DELETE FROM tree WHERE lft >= $lft AND rgt <= $rgt
-        connection.execute(
-            table.delete(and_(table.c.lft >= lft, table.c.rgt <= rgt,
-                              table.c.tree_id == tree_id))
-        )
+        if delete:
+            connection.execute(
+                table.delete(and_(table.c.lft >= lft, table.c.rgt <= rgt,
+                                  table.c.tree_id == tree_id))
+            )
 
         if instance.parent_id:
             """ Update key of current tree
@@ -111,24 +112,118 @@ class NestedSetsExtension(object):
         """ step 0: Initialize parameters.
         """
         # put there id of moving node
-        node_id = self.id
+        node_id = instance.id
 
         # put there left and right position of moving node
-        node_pos_left, node_pos_right = connection.execute(
-            select([table.c.lft, table.c.rgt])
+        (node_pos_left, node_pos_right,
+         node_tree_id, node_parent_id) = connection.execute(
+            select([table.c.lft, table.c.rgt,
+                    table.c.tree_id, table.c.parent_id])
             .where(table.c.id == node_id)
         ).fetchone()
 
         # put there id of new parent node (there moving node should be moved)
         # put there right position of new parent node (there moving node should be moved)
-        parent_id, parent_pos_right = connection.execute(
-            select([table.c.id, table.c.rgt])
+        (parent_id, parent_pos_right,
+         parent_pos_left, parent_tree_id) = connection.execute(
+            select([table.c.id, table.c.rgt, table.c.lft, table.c.tree_id])
             .where(table.c.id == instance.parent_id)
         ).fetchone()
 
         # 'size' of moving node (including all it's sub nodes)
         node_size = node_pos_right - node_pos_left + 1
 
+        if node_tree_id == parent_tree_id:
+            """ If move node in current tree
+            """
+            self._update_tree(table, connection, node_id, node_size,
+                              node_pos_left, node_pos_right, parent_pos_right,
+                              node_tree_id)
+        elif node_tree_id != parent_tree_id:
+            """ If move node to another tree
+
+                old_instance = connection.execute(
+                    select([table])
+                    .where(table.c.id == node_id)
+                ).fetchone()
+                from sqlalchemy.orm import sessionmaker
+                session = sessionmaker(bind=connection.engine)
+                DBSession = session()
+                old_instance = DBSession.query(mapper.class_).\
+                    filter(mapper.class_.id == node_id).one()
+            """
+            """ get subtree from node
+
+                SELECT id, name, level FROM my_tree
+                WHERE left_key >= $left_key AND right_key <= $right_key
+                ORDER BY left_key
+            """
+            subtree = connection.execute(
+                select([table.c.id])
+                .where(and_(table.c.lft >= node_pos_left,
+                            table.c.rgt <= node_pos_right,
+                            table.c.tree_id == node_tree_id))
+                .order_by(table.c.lft)
+            ).fetchall()
+            subtree = map(lambda x: x[0], subtree)
+
+            # new parent right
+            delta = node_size * 2 + node_pos_left
+
+            """ delete from old tree
+            """
+            self.mptt_after_delete(mapper, connection, instance, False)
+
+            """ insert subtree in exist tree
+            """
+            self._insert_subtree(table, connection, node_id, node_size,
+                                 node_pos_left, node_pos_right, parent_pos_left,
+                                 parent_pos_right, node_tree_id, subtree,
+                                 parent_tree_id, delta)
+            instance.tree_id = parent_tree_id
+            instance.right = delta
+
+    @classmethod
+    def _insert_subtree(self, table, connection, node_id, node_size,
+                        node_pos_left, node_pos_right, parent_pos_left,
+                        parent_pos_right, node_tree_id, subtree,
+                        parent_tree_id, delta):
+        """ step 1: rebuild inserted subtree
+        """
+        connection.execute(
+            table.update(table.c.id.in_(subtree))
+            .values(
+                lft=table.c.lft-node_pos_left+parent_pos_left+1,
+                rgt=table.c.rgt-node_pos_right+delta,
+                tree_id=parent_tree_id
+            )
+        )
+
+        """ step2: rebuild parent branch
+
+            SELECT id, name, level FROM my_tree
+            WHERE left_key <= $left_key AND right_key >= $right_key
+            ORDER BY left_key
+        """
+        parent_branch = connection.execute(
+            select([table.c.id])
+            .where(and_(table.c.lft <= parent_pos_left,
+                        table.c.rgt >= parent_pos_right,
+                        table.c.tree_id == parent_tree_id))
+            .order_by(table.c.lft)
+        ).fetchall()
+        parent_branch = map(lambda x: x[0], parent_branch)
+
+        connection.execute(
+            table.update(table.c.id.in_(parent_branch))
+            .values(
+                rgt=table.c.rgt-parent_pos_right+delta+1,
+            )
+        )
+
+    @classmethod
+    def _update_tree(self, table, connection, node_id, node_size, node_pos_left,
+                     node_pos_right, parent_pos_right, node_tree_id):
         """ step 1: temporary "remove" moving node
 
             UPDATE `list_items`
@@ -138,7 +233,8 @@ class NestedSetsExtension(object):
         connection.execute(
             table.update(
                 and_(table.c.lft >= node_pos_left,
-                     table.c.rgt <= node_pos_right))
+                     table.c.rgt <= node_pos_right,
+                     table.c.tree_id == node_tree_id))
             .values(
                 lft=0-table.c.lft,
                 rgt=0-table.c.rgt,
@@ -157,11 +253,13 @@ class NestedSetsExtension(object):
             WHERE `pos_right` > @node_pos_right;
         """
         connection.execute(
-            table.update(table.c.lft > node_pos_right)
+            table.update(and_(table.c.lft > node_pos_right,
+                         table.c.tree_id == node_tree_id))
             .values(lft=table.c.lft-node_size)
         )
         connection.execute(
-            table.update(table.c.rgt > node_pos_right)
+            table.update(and_(table.c.rgt > node_pos_right,
+                         table.c.tree_id == node_tree_id))
             .values(rgt=table.c.rgt-node_size)
         )
 
@@ -178,7 +276,8 @@ class NestedSetsExtension(object):
             clause = parent_pos_right - node_size
 
         connection.execute(
-            table.update(table.c.lft >= clause)
+            table.update(and_(table.c.lft >= clause,
+                         table.c.tree_id == node_tree_id))
             .values(lft=table.c.lft+node_size)
         )
         """
@@ -188,7 +287,8 @@ class NestedSetsExtension(object):
                         @parent_pos_right - @node_size, @parent_pos_right);
         """
         connection.execute(
-            table.update(table.c.rgt >= clause)
+            table.update(and_(table.c.rgt >= clause,
+                         table.c.tree_id == node_tree_id))
             .values(rgt=table.c.rgt+node_size)
         )
 
@@ -210,7 +310,8 @@ class NestedSetsExtension(object):
 
         connection.execute(
             table.update(and_(table.c.lft <= 0-node_pos_left,
-                              table.c.rgt >= 0-node_pos_right))
+                              table.c.rgt >= 0-node_pos_right,
+                              table.c.tree_id == node_tree_id))
             .values(lft=0-table.c.lft+clause,
                     rgt=0-table.c.rgt+clause)
         )
